+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased] - (release date)
 
+## [1.8.1] - 2026-06-14
+
 ### Changed
 
 - Queries longer than 400 characters are now supported.
@@ -176,7 +178,8 @@
 Initial version
 
 <!-- next-url -->
-[Unreleased]: https://github.com/matthias-stemmler/annimate/compare/v1.8.0...HEAD
+[Unreleased]: https://github.com/matthias-stemmler/annimate/compare/v1.8.1...HEAD
+[1.8.1]: https://github.com/matthias-stemmler/annimate/compare/v1.8.0...v1.8.1
 [1.8.0]: https://github.com/matthias-stemmler/annimate/compare/v1.7.0...v1.8.0
 [1.7.0]: https://github.com/matthias-stemmler/annimate/compare/v1.6.0...v1.7.0
 [1.6.0]: https://github.com/matthias-stemmler/annimate/compare/v1.5.0...v1.6.0